Task 6 perevoschikov

+++ src/main/java/org/example/Storage.java
@@ -2,18 +2,32 @@ package org.example;
 import java.sql.*;
 import java.util.ArrayList;
 
+/**
+ * Класс для управления хранилищем книг и цитат.
+ */
 public class Storage implements BookStorage {
 
-    private static final String DATABASE_URL = "jdbc:sqlite:D:\\Task5\\read_books.db";
 
+    /**
+     * URL базы данных SQLite.
+     */
+    private static final String DATABASE_URL = "jdbc:sqlite:read_books.db";
+
+
+    /**
+     * Соединение с базой данных.
+     */
     private Connection connection;
 
 
+    /**
+     * Список цитат.
+     */
     final private ArrayList<String> quoteList;
 
 
     /**
-     * Хранилище для цитат
+     * Конструктор класса. Инициализирует соединение с базой данных и список цитат.
      */
     public Storage() {
         try {
@@ -77,26 +91,6 @@ public class Storage implements BookStorage {
      * Метод для замены книги в списке прочитанных книг по формату: старое_название /n старый_автор /n старый_год новое_название /n новый_автор /n новый_год
      */
     public void editReadBook(String oldTitle, String oldAuthor, int oldYear, String newTitle, String newAuthor, int newYear, long chatId) {
-        try (PreparedStatement statement = connection.prepareStatement(
-                "UPDATE read_books SET title = ?, author = ?, year = ? WHERE title = ? AND author = ? AND year = ? AND chat_id = ?")) {
-            statement.setString(1, newTitle);
-            statement.setString(2, newAuthor);
-            statement.setInt(3, newYear);
-            statement.setString(4, oldTitle);
-            statement.setString(5, oldAuthor);
-            statement.setInt(6, oldYear);
-            statement.setLong(7, chatId);
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-
-    /**
-     * Метод для замены книги в списке рекомендованных книг
-     */
-    public void editRecBook(String oldTitle, String oldAuthor, int oldYear, String newTitle, String newAuthor, int newYear, long chatId) {
         try (PreparedStatement statement = connection.prepareStatement(
                 "UPDATE read_books SET title = ?, author = ?, year = ? WHERE title = ? AND author = ? AND year = ? AND chat_id = ?")) {
             statement.setString(1, newTitle);
@@ -190,49 +184,14 @@ public class Storage implements BookStorage {
     /**
      * Метод для обновления списка прочитанных книг
      */
-    public void updateReadBooks(long chatId, ArrayList<String> readBooks) {
-        try {
-            connection.setAutoCommit(false);
-
-            ArrayList<String> currentBooks = getReadBooks(chatId);
-
-            for (String book : currentBooks) {
-                if (!readBooks.contains(book)) {
-                    try (PreparedStatement deleteStatement = connection.prepareStatement(
-                            "DELETE FROM read_books WHERE chat_id = ? AND title = ?")) {
-                        deleteStatement.setLong(1, chatId);
-                        deleteStatement.setString(2, book);
-                        deleteStatement.executeUpdate();
-                    }
-                }
-            }
-
-            try (PreparedStatement selectStatement = connection.prepareStatement(
-                    "SELECT * FROM read_books WHERE chat_id = ? AND title = ?")) {
-                for (String book : readBooks) {
-                    selectStatement.setLong(1, chatId);
-                    selectStatement.setString(2, book);
-                    try (ResultSet resultSet = selectStatement.executeQuery()) {
-                        if (!resultSet.next()) {
-                            try (PreparedStatement insertStatement = connection.prepareStatement(
-                                    "INSERT INTO read_books (title, chat_id) VALUES (?, ?)")) {
-                                insertStatement.setString(1, book);
-                                insertStatement.setLong(2, chatId);
-                                insertStatement.executeUpdate();
-                            }
-                        }
-                    }
-                }
-            }
-
-            connection.commit();
-            connection.setAutoCommit(true);
-        } catch (Exception e) {
-            try {
-                connection.rollback();
-            } catch (SQLException ex) {
-                ex.printStackTrace();
-            }
+    public void updateReadBooks(long chatId, String oldTitle, String oldAuthor, int oldYear) {
+        try (PreparedStatement statement = connection.prepareStatement("DELETE FROM read_books WHERE chat_id = ? AND title = ? AND author = ? AND year = ?")) {
+            statement.setLong(1, chatId);
+            statement.setString(2, oldTitle);
+            statement.setString(3, oldAuthor);
+            statement.setInt(4, oldYear);
+            statement.executeUpdate();
+        } catch (SQLException e) {
             e.printStackTrace();
         }
     }
@@ -260,7 +219,14 @@ public class Storage implements BookStorage {
     }
 
 
-
+    /**
+     * Метод для добавления рекомендованной книги в базу данных.
+     *
+     * @param title   Название книги.
+     * @param author  Автор книги.
+     * @param genre   Жанр книги.
+     * @param chatId  Идентификатор чата.
+     */
     public void addRecBook(String title, String author, String genre, long chatId) {
         try (PreparedStatement statement = connection.prepareStatement("INSERT INTO recommendedBooks (title, author, genre, chat_id) VALUES (?, ?, ?, ?)")) {
             statement.setString(1, title);
@@ -273,6 +239,12 @@ public class Storage implements BookStorage {
         }
     }
 
+
+    /**
+     * Метод для получения списка названий рекомендованных книг.
+     *
+     * @return Список названий книг.
+     */
     public ArrayList<String> getRecBooks() {
         ArrayList<String> books = new ArrayList<>();
         try (PreparedStatement statement = connection.prepareStatement("SELECT title FROM recommendedBooks")) {
@@ -288,14 +260,20 @@ public class Storage implements BookStorage {
     }
 
 
-    public boolean recBookExists(String title, String author, String genre, long chatId) {
+
+    /**
+     * Метод для проверки существования рекомендованной книги.
+     *
+     * @param title  Название книги.
+     * @param author Автор книги.
+     * @return {@code true}, если книга существует, иначе {@code false}.
+     */
+    public boolean recBookExists(String title, String author) {
         boolean exists = false;
         try (PreparedStatement statement = connection.prepareStatement(
-                "SELECT * FROM recommendedBooks WHERE title = ? AND author = ? AND genre = ? AND chat_id = ?")) {
+                "SELECT * FROM recommendedBooks WHERE title = ? AND author = ?")) {
             statement.setString(1, title);
             statement.setString(2, author);
-            statement.setString(3, genre);
-            statement.setLong(4, chatId);
             try (ResultSet resultSet = statement.executeQuery()) {
                 exists = resultSet.next();
             }
@@ -305,6 +283,13 @@ public class Storage implements BookStorage {
         return exists;
     }
 
+
+    /**
+     * Метод для поиска книг по указанному жанру.
+     *
+     * @param genre Жанр книг.
+     * @return Список найденных книг в формате "название от автора".
+     */
     public ArrayList<String> searchBooksByGenre(String genre) {
         ArrayList<String> books = new ArrayList<>();
         try (PreparedStatement statement = connection.prepareStatement("SELECT title, author FROM recommendedBooks WHERE genre = ?")) {
@@ -322,6 +307,13 @@ public class Storage implements BookStorage {
         return books;
     }
 
+
+    /**
+     * Метод для поиска книг по указанному автору.
+     *
+     * @param author Автор книг.
+     * @return Список найденных книг в формате "название (жанр)".
+     */
     public ArrayList<String> searchBooksByAuthor(String author) {
         ArrayList<String> books = new ArrayList<>();
         try (PreparedStatement statement = connection.prepareStatement("SELECT title, genre FROM recommendedBooks WHERE author = ?")) {
@@ -340,9 +332,54 @@ public class Storage implements BookStorage {
     }
 
 
+    /**
+     * Метод для обновления данных о рекомендованных книгах.
+     *
+     * @param chatId   Идентификатор чата.
+     * @param oldTitle Старое название книги.
+     * @param oldAuthor Старый автор книги.
+     * @param oldGenre  Старый жанр книги.
+     */
+    public void updateRecBooks(long chatId, String oldTitle, String oldAuthor, String oldGenre) {
+        try (PreparedStatement statement = connection.prepareStatement("DELETE FROM recommendedBooks WHERE chat_id = ? AND title = ? AND author = ? AND genre = ?")) {
+            statement.setLong(1, chatId);
+            statement.setString(2, oldTitle);
+            statement.setString(3, oldAuthor);
+            statement.setString(4, oldGenre);
+            statement.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
 
 
+    /**
+     * Метод для получения всех значений о рекомендованных книгах.
+     *
+     * @return Список строк с данными о книгах.
+     */
+    public ArrayList<String> getAllRecValues() {
+        ArrayList<String> allValues = new ArrayList<>();
+        try (PreparedStatement statement = connection.prepareStatement("SELECT title, author, genre, chat_id FROM recommendedBooks")) {
+            try (ResultSet resultSet = statement.executeQuery()) {
+                while (resultSet.next()) {
+                    String title = resultSet.getString("title");
+                    String author = resultSet.getString("author");
+                    String genre = resultSet.getString("genre");
+                    long chat_id = resultSet.getLong("chat_id");
+                    allValues.add(title + "\n" + author + "\n" + genre +"\n" + chat_id);
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return allValues;
+    }
 
+
+    /**
+     * Метод для закрытия соединения с базой данных.
+     */
     public void closeConnection() {
         try {
             if (connection != null) {
